Simplify item loading and saving using plugins
ItemLoaderInterface methods canLoadItems() and canSaveItems() determine
if tab can be loaded and saved by a plugin. If method loadItems() or
saveItems() return false, the tab is disabled.

+++ src/gui/configurationmanager.cpp
@@ -59,6 +59,23 @@ void printItemFileError(const QString &id, const QString &fileName, const QFile 
          , LogError );
 }
 
+bool needToSaveItemsAgain(const QAbstractItemModel &model, const ItemFactory &itemFactory,
+                          const ItemLoaderInterfacePtr &currentLoader)
+{
+    if (!currentLoader)
+        return false;
+
+    bool saveWithCurrent = true;
+    foreach ( const ItemLoaderInterfacePtr &loader, itemFactory.loaders() ) {
+        if ( itemFactory.isLoaderEnabled(loader) && loader->canSaveItems(model) )
+            return loader != currentLoader;
+        else if (loader == currentLoader)
+            saveWithCurrent = false;
+    }
+
+    return !saveWithCurrent;
+}
+
 } // namespace
 
 // singleton
@@ -117,43 +134,47 @@ ItemLoaderInterfacePtr ConfigurationManager::loadItems(ClipboardModel &model)
 
     ItemLoaderInterfacePtr loader;
 
+    model.setDisabled(true);
+
     if ( file.exists() ) {
         COPYQ_LOG( QString("Tab \"%1\": Loading items").arg(tabName) );
         if ( file.open(QIODevice::ReadOnly) )
             loader = itemFactory()->loadItems(&model, &file);
-        else
+
+        if ( needToSaveItemsAgain(model, *itemFactory(), loader) ) {
+            COPYQ_LOG( QString("Tab \"%1\": Saving items using other plugin").arg(tabName) );
             model.setDisabled(true);
+            loader.clear();
+            file.close();
+            if ( file.open(QIODevice::WriteOnly) ) {
+                loader = itemFactory()->initializeTab(&model);
+                saveItems(model, loader);
+            } else {
+                COPYQ_LOG( QString("Tab \"%1\": Failed to re-save items").arg(tabName) );
+            }
+        }
     } else {
         COPYQ_LOG( QString("Tab \"%1\": Creating new tab").arg(tabName) );
-        if ( file.open(QIODevice::ReadWrite) )
-            loader = itemFactory()->createTab(&model, &file);
-        else
-            model.setDisabled(true);
-    }
-
-    COPYQ_LOG( QString("Tab \"%1\": %2 items loaded").arg(tabName).arg(model.rowCount()) );
-
-    if ( model.isDisabled() ) {
-        COPYQ_LOG( QString("Tab \"%1\": Disabled").arg(tabName) );
-        return ItemLoaderInterfacePtr();
+        if ( file.open(QIODevice::WriteOnly) ) {
+            loader = itemFactory()->initializeTab(&model);
+            saveItems(model, loader);
+        }
     }
 
     file.close();
-    file.open(QIODevice::ReadOnly);
-    itemFactory()->itemsLoaded(&model, &file);
 
-    if ( model.isDirty() ) {
-        COPYQ_LOG( QString("Tab \"%1\": Dirty").arg(tabName) );
-        ItemLoaderInterfacePtr loader2 = saveItems(model);
-        if ( !loader2.isNull() )
-            loader = loader2;
-        model.setDirty(false);
-    }
+    if (loader)
+        COPYQ_LOG( QString("Tab \"%1\": %2 items loaded").arg(tabName).arg(model.rowCount()) );
+    else
+        COPYQ_LOG( QString("Tab \"%1\": Disabled").arg(tabName) );
+
+    model.setDisabled(!loader);
 
     return loader;
 }
 
-ItemLoaderInterfacePtr ConfigurationManager::saveItems(const ClipboardModel &model)
+ItemLoaderInterfacePtr ConfigurationManager::saveItems(const ClipboardModel &model,
+                                                       const ItemLoaderInterfacePtr &loader)
 {
     const QString tabName = model.property("tabName").toString();
     const QString fileName = itemFileName(tabName);
@@ -170,14 +191,16 @@ ItemLoaderInterfacePtr ConfigurationManager::saveItems(const ClipboardModel &mod
 
     COPYQ_LOG( QString("Tab \"%1\": Saving %2 items").arg(tabName).arg(model.rowCount()) );
 
-    ItemLoaderInterfacePtr loader = itemFactory()->saveItems(model, &file);
+    if ( loader->saveItems(model, &file) ) {
+        COPYQ_LOG( QString("Tab \"%1\": Items saved").arg(tabName) );
 
-    COPYQ_LOG( QString("Tab \"%1\": Items saved").arg(tabName) );
-
-    // Overwrite previous file.
-    QFile::remove(fileName);
-    if ( !file.rename(fileName) )
-        printItemFileError(tabName, fileName, file);
+        // Overwrite previous file.
+        QFile::remove(fileName);
+        if ( !file.rename(fileName) )
+            printItemFileError(tabName, fileName, file);
+    } else {
+        COPYQ_LOG( QString("Tab \"%1\": Failed to save items!").arg(tabName) );
+    }
 
     return loader;
 }
